move bp_history to own file
