+++ hypertension/dc-7101-algorithms/py-root/lib/utils.py
@@ -103,24 +103,6 @@ def bp_readings_meet_date_specs(date_of_claim, bp_readings):
     return claim_within_one_month and claim_within_six_months
 
 
-def history_of_diastolic_bp(request):
-    diastolic_history_calculation = {
-        "success": True
-    }
-    bp_readings = request["bp"]
-    bp_readings_length = len(bp_readings)
-    readings_greater_or_equal_to_one_hundred = 0
-
-    if bp_readings_length > 0:
-        for reading in bp_readings:
-            if reading["diastolic"] >= 100:
-                readings_greater_or_equal_to_one_hundred += 1
-        diastolic_history_calculation["diastolic_bp_predominantly_100_or_more"] = True if readings_greater_or_equal_to_one_hundred / bp_readings_length >=.5 else False
-    else:
-        diastolic_history_calculation["success"] = False
-
-    return diastolic_history_calculation
-
 def sufficient_to_autopopulate (request):
     predominance_calculation = {
         "success": True,
